Revert Debian to Bookworm
There are currently several issues with Debian (and other distro's) using MariaDB.
Since there is no clean way to use Oracle's MySQL packages we are currently forced to stop building Debian containers.

Also reverting to Bookworm is not an option as it happens with the latest Bookworm too.

This commit will build Alpine as the one and only container image, and will use the same tags as the Debian builds did.
For more details read:
- https://github.com/dani-garcia/vaultwarden/issues/6416
- https://github.com/diesel-rs/diesel/issues/4699
- https://github.com/diesel-rs/diesel/discussions/4806
- https://jira.mariadb.org/browse/CONC-782

Fixes #6416

Signed-off-by: BlackDex <black.dex@gmail.com>